splits tests further
